+++ .gitignore
@@ -1,3 +1,4 @@
 .DS_Store
 node_modules
 dist
+junit.xml
