+++ src/index.ts
@@ -1,36 +1,2 @@
-import marked, { MarkedOptions } from 'marked';
-import type { PluginImpl } from 'rollup';
-import { createFilter } from 'rollup-pluginutils';
-
-const ext = /\.md$/;
-
-export type Options = { marked?: MarkedOptions } & {
-  include?: string | RegExp | (string | RegExp)[];
-  exclude?: string | RegExp | (string | RegExp)[];
-};
-
-export const md: PluginImpl<Options> = (options = {}) => {
-  const filter = createFilter(options.include || ['**/*.md'], options.exclude);
-  if (options.marked) {
-    marked.setOptions(options.marked);
-  }
-  return {
-    name: 'md',
-
-    transform(md, id) {
-      if (!ext.test(id)) {
-        return null;
-      }
-
-      if (!filter(id)) {
-        return null;
-      }
-
-      const data = marked(md);
-      return {
-        code: `export default ${JSON.stringify(data.toString())};`,
-        map: { mappings: '' },
-      };
-    },
-  };
-};
+export type { MdOptions } from './md';
+export { md } from './md';
