+++ src/samples/main.js
@@ -1,0 +1,2 @@
+import md from './test.md';
+console.log('md: ', md);
