--- test/samples/main.js
@@ -1,2 +1,0 @@
-import md from './test.md'
-console.log('md: ', md)
