+++ tsconfig.json
@@ -11,6 +11,7 @@
   "exclude": [
     "../../node_modules",
     "./dist",
-    "src/**/*.test.ts"
+    "src/**/*.test.ts",
+    "samples/**/*.*"
   ]
 }
